+++ packages/form/lib/src/form.dart
@@ -2,3 +2,4 @@ library form;
 
 export 'form/bond_form_controller.dart';
 export 'form/bond_form_state.dart';
+export 'form/field_change_tracking_mixin.dart';