+++ packages/form/packages/bond_form_riverpod/lib/src/form/auto_disposal_form_state_notifier.dart
@@ -17,6 +17,7 @@ abstract class AutoDisposeFormStateNotifier<Success, Failure extends Error>
     extends AutoDisposeNotifier<BondFormState<Success, Failure>>
     with
         BaseFormController<Success, Failure, BondFormState<Success, Failure>>,
+        FieldChangeTrackingMixin<Success, Failure>,
         FormController<Success, Failure> {
   /// Determines whether the validation should stop at the first error.
   final bool stopOnFirstError;