+++ packages/form/packages/bond_form_riverpod/lib/src/form/form_state_notifier.dart
@@ -13,6 +13,7 @@ abstract class FormStateNotifier<Success, Failure extends Error>
     extends Notifier<BondFormState<Success, Failure>>
     with
         BaseFormController<Success, Failure, BondFormState<Success, Failure>>,
+        FieldChangeTrackingMixin<Success, Failure>,
         FormController<Success, Failure> {
   /// Determines whether the validation should stop at the first error.
   final bool stopOnFirstError;